Privileges support.
There is a new option 'privileges' to control the access privileges
output. The default is false.

It is a generic infrastructure that is capable of output GRANT and
REVOKE commands for a bunch of objects. This commit support domain,
sequence, table, function, schema, language, and type. Some tests were
added too.

Column privileges are not supported yet.