fix reloptions output for various objects
When both tables have reloptions, the diff SQL command was not correct
due to a bug in intersectWithSortedList. We don't fix the function.
Instead, we add a conservative solution that consists of output all
options from the table in 'to'. It is not wrong, but it would be nice to
fix it someday to omit option/value that doesn't change.

Also, add a small test to catch similar bugs.