snprintf() is not portable
Windows has a different behaviour for snprintf function. It does not
return the size of a formatted string if we have to truncate the result
string. Instead, we rely on the following behaviour:

r = snprintf(NULL, 0, "%s", buf);

r will contain the size of buf. After that we allocate r + 1 bytes and
call snprintf again to fill query variable.