Rename statistics parameter
Change parameter 'statistics' to 'summary' because there is a new
PostgreSQL command named STATISTICS. This parameter was not added to
test configuration files. I assume that it is not widely used because
there is not a documentation. I intend to use this parameter name
(statistics) for another commit (that implements STATISTICS command).