Support partitioning syntax
Add following commands:

CREATE TABLE ... PARTITION BY
ALTER TABLE ... ATTACH PARTITION
ALTER TABLE ... DETACH PARTITION

It means that it is possible to create partitioned tables and
attach/detach partitions from a partitioned table.

However, it is not possible to transform a regular table into
partitioned table and vice-versa (Postgres does not have a syntax to
it. We could drop and recreate the table but it could lost data).