Add command-line options to each configuration file option
Since configuration file is optional (commit c13e82f), provide options
on the command line. It is easier for scripts to use command-line
options rather than write a file with a bunch of options.
I also added some options that was already supported on the command-line
to configuration file (single-transaction, ignore-version, and
temp-directory). BTW, I renamed option tmpdir to temp-directory.