Prompt for password
pgquarrel will automatically prompt for a password if the server demands
password authentication. However, pgquarrel will waste a connection
attempt finding out that the server wants a password. This behavior is
equivalent to --password option (psql).

A password prompt will appear if a password is not available by other
means such as configuration file or .pgpass file.