PostgreSQL 12 support
It is a good time to support a new PostgreSQL version since it is in
beta time.
I decided to use more PostgreSQL code instead of duplicate code in
pgquarrel. ScanKeywordLookup() uses backend function when available (>=
9.6). Also, ScanKeywordLookup signature changes in v12. Since some of
the backend functions used from now on is in libpgcommon, optionally add
it -- some ancient versions (<= 9.2) does not have libpgcommon.