Avoid some Debian/Ubuntu warnings
If you have multiple PostgreSQL versions, Debian/Ubuntu install a copy
of some include files from the latest(?) PostgreSQL version into
/usr/include/postgresql. It causes some warnings (if you are compiling
with a version that is not the latest) because there are some specific
include files into /usr/include/postgresql/XX too. The workaround is to
include the specific include files first.