Policy bugs
It fixes a possible dereference in options withcheck, qual, and roles
while calling strcmp. There is also two bugs: (i) forgot to treat NULL
in case command is to remove WITH CHECK expression, USING expression,
and roles and (ii) forgot to add a semicolon at the end of SQL command.