Refactor language
Since pg_pltemplate was removed in version 13, a refactor was done to
remove its usage. Even in prior versions, we can build a query without
using pg_pltemplate because we can test the existence of handler, inline
and validator functions. While on it, I added schema for each function
because it can be in a different schema.