test: make PG directories configurable, PGQUARREL path configurable
This allows for the directory to the PostgreSQL installation to be
configurable using the "PGPATH_<VERSION>" environment variable. For
example, if running tests using `run-test.sh 14 14` and you have
PGPATH_14 in your env, it'll use that instead of $HOME/pg14.

This also allows for configuring the path to the `pgquarrel` binary by
setting the PGQUARREL environment variable.

Finally, we change the unix socket directory to the temp directory so
that we don't need permission to /var/run or /run (whatever the default
is for your system) since this isn't always there.